refactor: split HostnameInput into presentational/container components
